User: Reorganize auth view tests
